fixed differences

+++ src/searchparams.rs
@@ -9,8 +9,8 @@ pub const RFP_MARGIN_FACTOR: i32 = 1200;
 
 pub const AFP_MARGIN: i32 = 30000;
 
-pub const NULL_MOVE_R_BASE: f32 = 4.0;
-pub const NULL_MOVE_R_FACTOR: f32 = 1.0 / 6.0;
+pub const NULL_MOVE_R_BASE: f64 = 4.0;
+pub const NULL_MOVE_R_FACTOR: f64 = 1.0 / 6.0;
 pub const NULL_MOVE_R_DENOM: i32 = 3000;
 
 pub const RAZORING_MARGIN: i32 = 2500;
@@ -19,13 +19,13 @@ pub const LMR_BASE: f64 = 0.8;
 pub const LMR_FACTOR: f64 = 1.0 / 2.25;
 pub const LMR_HISTORY_DENOMINATOR: i32 = 8000;
 
-pub const LMP_IMPROVING_BASE: f32 = 4.0;
-pub const LMP_IMPROVING_FACTOR: f32 = 1.0;
-pub const LMP_NONIMPROVING_BASE: f32 = 2.0;
-pub const LMP_NONIMPROVING_FACTOR: f32 = 0.5;
+pub const LMP_IMPROVING_BASE: f64 = 4.0;
+pub const LMP_IMPROVING_FACTOR: f64 = 1.0;
+pub const LMP_NONIMPROVING_BASE: f64 = 2.0;
+pub const LMP_NONIMPROVING_FACTOR: f64 = 0.5;
 
-pub const HISTORY_DECAY_FACTOR: f32 = 1.0 / 512.0;
-pub const HISTORY_DELTA_FACTOR: f32 = 32.0;
+pub const HISTORY_DECAY_FACTOR: f64 = 1.0 / 512.0;
+pub const HISTORY_DELTA_FACTOR: f64 = 32.0;
 
 pub const HISTORY_LEAF_PRUNING_MARGIN: i32 = 6000;
 pub const COUNTERMOVE_PRUNING_FACTOR: i32 = -700;
